+++ internal/versions/versions.go
@@ -1,6 +1,0 @@
-package versions
-
-const (
-	SourceProtocolVersion      = uint64(1)
-	DestinationProtocolVersion = uint64(1)
-)