+++ serve/destination_test.go
@@ -34,14 +34,25 @@ func (*testDestinationClient) Initialize(context.Context, specs.Destination) err
 func (*testDestinationClient) Migrate(context.Context, schema.Tables) error {
 	return nil
 }
-func (*testDestinationClient) Write(context.Context, string, map[string]interface{}) error {
+
+func (*testDestinationClient) Read(context.Context, *schema.Table, string, chan<- *schema.DestinationResource) error {
 	return nil
+}
+func (*testDestinationClient) Write(_ context.Context, _ schema.Tables, resources <-chan *schema.DestinationResource) error {
+	//nolint:revive
+	for range resources {
+	}
+	return nil
+}
+
+func (*testDestinationClient) Metrics() plugins.DestinationMetrics {
+	return plugins.DestinationMetrics{}
 }
 
 func (*testDestinationClient) Close(context.Context) error {
 	return nil
 }
-func (*testDestinationClient) DeleteStale(context.Context, string, string, time.Time) error {
+func (*testDestinationClient) DeleteStale(context.Context, schema.Tables, string, time.Time) error {
 	return nil
 }
 
@@ -50,7 +61,6 @@ func TestDestination(t *testing.T) {
 	s := &destinationServe{
 		plugin: plugin,
 	}
-
 	cmd := newCmdDestinationRoot(s)
 	cmd.SetArgs([]string{"serve", "--network", "test"})
 	ctx := context.Background()
@@ -111,29 +121,34 @@ func TestDestination(t *testing.T) {
 		t.Fatalf("expected version to be development but got %s", version)
 	}
 
-	if err := c.Migrate(ctx, schema.Tables{testTable()}); err != nil {
+	tables := schema.Tables{testTable()}
+	if err := c.Migrate(ctx, tables); err != nil {
 		t.Fatal(err)
 	}
 
 	resource := schema.NewResourceData(testTable(), nil, nil)
-	resource.Data["id"] = "test"
-	b, err := json.Marshal(resource)
+	if err := resource.Set("test_column", 5); err != nil {
+		t.Fatal(err)
+	}
+	destResource := resource.ToDestinationResource()
+	b, err := json.Marshal(destResource)
 	if err != nil {
 		t.Fatal(err)
 	}
 	resources := make(chan []byte, 1)
 	resources <- b
 	close(resources)
-	failedWrites, err := c.Write(ctx, "test", time.Now(), resources)
-	if err != nil {
+	if err := c.Write2(ctx, tables, "test", time.Now(), resources); err != nil {
 		t.Fatal(err)
-	}
-	if failedWrites != 0 {
-		t.Fatalf("expected failed writes to be 0 but got %d", failedWrites)
 	}
 
 	if err := c.DeleteStale(ctx, nil, "testSource", time.Now()); err != nil {
 		t.Fatalf("failed to call DeleteStale: %v", err)
+	}
+
+	_, err = c.GetMetrics(ctx)
+	if err != nil {
+		t.Fatal(err)
 	}
 
 	if err := c.Close(ctx); err != nil {